requirements.txt
Deprecated==1.2.18
Jinja2==3.1.6
accelerate
airportsdata==20250523
astor==0.8.1
blake3==1.0.5
compressed-tensors==0.9.3
datasets
depyf==0.18.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
gguf==0.17.1
googleapis-common-protos==1.70.0
hf-xet==1.1.4
httptools==0.6.4
huggingface-hub==0.33.0
importlib_metadata==8.0.0
interegular==0.3.3
jiter==0.10.0
lark==1.2.2
llguidance==0.7.29
llvmlite==0.44.0
lm-format-enforcer==0.10.11
mistral_common==1.6.2
msgspec==0.19.0
numba==0.61.2
openai==1.88.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.9
outlines==0.1.11
outlines_core==0.1.26
partial-json-parser==0.2.1.1.post5
prometheus-fastapi-instrumentator==7.1.0
py-cpuinfo==9.0.0
pycountry==24.6.1
python-dotenv==1.1.0
python-multipart==0.0.20
ray==2.47.1
rich-toolkit==0.14.7
sentencepiece==0.2.0
setuptools==80.9.0
starlette==0.46.2
tiktoken==0.9.0
tokenizers==0.21.1
torchaudio==2.6.0
torchvision==0.21.0
transformers==4.52.4
uvicorn==0.34.3
uvloop==0.21.0
vllm==0.8.5
watchfiles==1.1.0
websockets==15.0.1
xformers==0.0.29.post2
xgrammar==0.1.18
